# PERF: use arr.size instead of np.prod(arr.shape) in _can_use_numexpr