--- a/pandas/core/array_algos/transforms.py
+++ b/pandas/core/array_algos/transforms.py
@@ -19,7 +19,7 @@ def shift(values: np.ndarray, periods: int, axis: int, fill_value) -> np.ndarray
         new_values = new_values.T
         axis = new_values.ndim - axis - 1
 
-    if np.prod(new_values.shape):
+    if new_values.size:
         new_values = np.roll(new_values, ensure_platform_int(periods), axis=axis)
 
     axis_indexer = [slice(None)] * values.ndim