@jbrockmendel 2021-02-15
> and I think it's always equivalent to .size

AFAIK this is correct

Could grep for this pattern elsewhere?

LGTM