@jorisvandenbossche 2021-02-16
Changed a few more (there are other cases, eg in nanops, but where we only have the shape, and not the original array, so that would require more code changes, and thus out of scope for this PR)